Add new Postman collections for Supermall API tests
- Introduced three new Postman collections: Supermall Auth API Tests, Supermall User Address API Tests, and Supermall User Profile API Tests.
- Each collection includes comprehensive test cases for user authentication, address management, and user profile management.
- Implemented various pre-request and test scripts to validate API responses and manage environment variables effectively.
- Enhanced organization and clarity of API testing for the Supermall e-commerce backend.

These additions improve the testing framework and ensure better coverage of critical functionalities.

+++ src/main/java/com/supermall/backend/domain/user/dto/UserFavoriteResponse.java
@@ -6,9 +6,9 @@ import java.time.LocalDateTime;
 
 @Data
 public class UserFavoriteResponse {
-    private Long id;
-    private Long userId;
-    private Long productId;
+    private Integer id;
+    private Integer userId;
+    private Integer productId;
     private LocalDateTime createdAt;
     
     // 商品信息
